Removed the unused enums and target type filter button class
